+++ userspace/lib/include/dirent.h
@@ -18,5 +18,10 @@ typedef struct DIR DIR;
 DIR *opendir(const char *path);
 struct dirent *readdir(DIR *dir);
 int closedir(DIR *dir);
+int scandir(const char *path,
+    struct dirent ***namelist,
+    int (*filter)(const struct dirent *),
+    int (*compar)(const struct dirent **, const struct dirent **));
+int alphasort(const struct dirent **left, const struct dirent **right);
 
 #endif