+++ userspace/lib/include/sys/stat.h
@@ -43,6 +43,7 @@ struct stat {
 };
 
 int stat(const char *path, struct stat *st);
+int lstat(const char *path, struct stat *st);
 int fstat(int fd, struct stat *st);
 int mkdir(const char *path, mode_t mode);
 int chmod(const char *path, mode_t mode);